chore(redirectFromDiscoveryToLoginClickUtility): Initial commit
