+++ src/Discovery/DiscoveryActionBar/components/DownloadManifestButton.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { FileTextOutlined } from '@ant-design/icons';
 import { Popover, Button } from 'antd';
 import handleDownloadManifestClick from '../utils/handleDownloadManifestClick';
-import handleRedirectToLoginClick from '../utils/handleRedirectToLoginClick';
+import { handleRedirectToLoginClick } from '../../Utils/HandleRedirectToLoginClick';
 
 const DownloadManifestButton = ({
   props, healIDPLoginNeeded, onlyInCommonMsg, history, location,
