+++ src/Discovery/DiscoveryActionBar/components/ExportToWorkspaceButton.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { ExportOutlined } from '@ant-design/icons';
 import { Popover, Button } from 'antd';
 import handleExportToWorkspaceClick from '../utils/handleExportToWorkspaceClick/handleExportToWorkspaceClick';
-import handleRedirectToLoginClick from '../utils/handleRedirectToLoginClick';
+import handleRedirectToLoginClick from '../../Utils/HandleRedirectToLoginClick';
 
 const ExportToWorkspaceButton = ({
   props, healIDPLoginNeeded, onlyInCommonMsg, setDownloadStatus, history, location,
